requirements.txt
pymssql
rouge
evaluate
nltk
python-dotenv